Extract underlay VNIC setup from opte to common::underlay
