+++ sled-agent/src/common/mod.rs
@@ -6,4 +6,5 @@
 
 pub mod disk;
 pub mod instance;
+pub mod underlay;
 pub mod vlan;
